% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.r
\name{build_area_point_grid}
\alias{build_area_point_grid}
\title{Build a point grid.}
\usage{
build_area_point_grid(area = NULL, data = NULL)
}
\description{
\code{build_area_p;oint_grid} builds a full rectangular point grid for an
 area.
}
\details{
@param area Character string.
 @param data Data frame. Should include all applicable areas in a "zone"
   column, along with "x" and "y" coordinates for each point in the zone.
 @return Returns data frame with full rectangle of points encompassing area.
   Points in grid that are included in LAI measurements are assigned their
   respective point.id. Points not included in LAI have point.id = NA.
}

