% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{build_auc_delta_stack}
\alias{build_auc_delta_stack}
\title{Build a stack of points and one year change in LAI area under curve.}
\usage{
build_auc_delta_stack(auc_stack = NULL, year_index = NULL)
}
\description{
\code{build_auc_delta_stack} builds a raster stack of the change in the LAI
 area under the curve from the previous year.
}
\details{
@param auc_stack Raster stack. A stack of rasters giving yearly AUC values.
   First layer in stack is point.id's, subsequent layers are AUC values.
 @param year_index Numeric vector. Vector of years for which AUC data is
   contained in auc_stack.
 @return Returns raster stack. First layer is point.id's, subsequent layers
   are % change in AUC from the previous year.
}

