% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{build_dates_index}
\alias{build_dates_index}
\title{Build an index of dates for which LAI images are available.}
\usage{
build_dates_index(data = owens_lai)
}
\description{
\code{build_dates_index} looks through LAI data in the format as supplied
 for Owens Lake and builds a list of all the dates for which LAI images
 were obtained.
}
\details{
@param data Data frame. LAI data.
 @return Returns a vector of dates.
}

