% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{build_lai_stack}
\alias{build_lai_stack}
\title{Build a stack of LAI values.}
\usage{
build_lai_stack(area = NULL)
}
\description{
\code{build_lai_stack} builds a raster stack of LAI images for a
 Seeps and Springs zone.
}
\details{
@param area Character string. Name of Seeps and Springs area.
 @return Returns raster stack. Layers have values = LAI reading from that
 image.
}

