% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{build_lai_window_summary}
\alias{build_lai_window_summary}
\title{Summarize LAI data around a day of the year.}
\usage{
build_lai_window_summary(data = NULL)
}
\description{
\code{bulid_lai_window_summary} summarizes the LAI data across multiple
 years in a month wide window around a normalized date.
}
\details{
@param data Data frame. Normalized dates in a column titled "normal date".
 @return Returns a data frame summarizing LAI data for available years of
   data for each day of the year.
}

