% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{build_points_stack}
\alias{build_points_stack}
\title{Build a stack of points and locations.}
\usage{
build_points_stack(area = NULL)
}
\description{
\code{build_points_stack} builds a raster stack of Landsat image points
 for a Seeps and Springs zone.
}
\details{
@param area Character string. Name of seeps and springs area.
 @return Returns raster stack with 2 layers. First layer value = point.id,
   second layer value is a number indicating the area if point in defined
   zone, 0 if outside.
}

