% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.r
\name{clean_lai}
\alias{clean_lai}
\title{Clean LAI dataframe.}
\usage{
clean_lai(data = NULL)
}
\description{
\code{clean_lai} cleans and checks a dataframe of lai values in the
 format as generated by \code{load_dbfs}.
}
\details{
@param data Dataframe. Dataframe of LAI values to be cleaned.
 @return Data frame with a point.id column and one column for each available
   LAI reading.
}

