% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.r
\name{clean_locs}
\alias{clean_locs}
\title{Clean point locations dataframe.}
\usage{
clean_locs(data = NULL)
}
\description{
\code{clean_points} cleans and checks the dataframe of point locations
 generated by \code{load_gdb_points}.
}
\details{
@param data Dataframe. Point locations object to be cleaned.
 @return Data frame with three columns: point.id, x.coord,
   and y.coord. Returns complete rows only.
}

