% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.r
\name{create_raster}
\alias{create_raster}
\title{Create raster.}
\usage{
create_raster(data = NULL)
}
\description{
\code{create_raster} builds a rater object out of a 3 column data frame.
}
\details{
@param data Data frame. A 3 column data frame with first two columns giving
   point coordinates and labeled "x" and "y". Third column will be raster
   point values.
 @return Returns raster of data with extents to encompass all points in
   data frame.
}

