% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai_functions.r
\name{get_patch_data}
\alias{get_patch_data}
\title{Build get data to patch missing offseason readings.}
\usage{
get_patch_data(data = NULL, patches = NULL)
}
\description{
\code{get_patch_data} averages LAI data from surrounding years to get data
 to replace missing data for a single pixel.
}
\details{
@param data Data frame. Vertical LAI data for the point.
 @param patches Data frame. Two column data frame: 1st column is list of
   dates where patch is needed, 2nd is those dates normalized by
   normalize_dates().
 @return Returns data frame with patch LAI values for required dates.
}

