% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.r
\name{isolate_area}
\alias{isolate_area}
\title{Strip a raster down to only those points contained within a Seeps and
 Springs zone.}
\usage{
isolate_area(zonelayer = NULL, valuelayer = NULL)
}
\description{
\code{isolate_area} compares a raster of values to a raster giving zone
 classification values and assigns a zero value to all points not contained
 in the main zone classification.
}
\details{
@param zonelayer Raster. A raster with cells assigned a numer corresponding
   to their seeps and springs zone, and NA if outside of an assigned zone.
 @param valuelayer Raster. A raster with values assigned to cells.
 @return Returns a raster with all points within the zone still holding
   their original value, and all points outside the zone assigned a value
   of NA.
}

