% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.r
\name{load_gdb_points}
\alias{load_gdb_points}
\title{Read geodatabase file.}
\usage{
load_gdb_points(file = NULL, layer = NULL)
}
\description{
\code{load_gdb_points} loads point locations from a GDB (geodatabase) file
 into a dataframe object.
}
\details{
@param file Character string. GDB file to be read, including path.
 @param layer Character string. Layer within GDB file to be read.
 @return Dataframe with three columns: point ID, x coordinate,
   and y coordinate
}

