% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.r
\name{melt_lai}
\alias{melt_lai}
\title{Melt LAI dataframe.}
\usage{
melt_lai(data = NULL)
}
\description{
\code{melt_lai} melts a wide dataframe of LAI values as generated by
 \code{load_dbfs} and and \code{clean_lai} into a melted data set. Cleans
 date values into proper format.
}
\details{
@param data Dataframe. Wide dataframe of LAI values.
 @return Data frame with a point.id column, a column for the date,
   and a column for the LAI value on the associated date.
}

