% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.r
\name{raster_to_df}
\alias{raster_to_df}
\title{Convert a raster to a data frame.}
\usage{
raster_to_df(seep_file = NULL, valuestack = NULL)
}
\description{
\code{raster_to_df} converts a raster stack of yearly values into a data
 in the format for sending to Chong (Formation).
}
\details{
@param seep_file Character string. The name of the discharge zone,
   lower case and _ for spaces.
 @param valuestack Character string. The stack object to data frame - ize.
 @return Writes the data frame to a .CSV file in the "output" subdirectory
   of the working directory.
}

