% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.r
\name{write_airscidb}
\alias{write_airscidb}
\title{Write to Air Sciences PostgreSQL database.}
\usage{
write_airscidb(data = NULL, tble = NULL)
}
\description{
\code{write_airscidb} writes from a dataframe into an existing table in the
 Air Sciences PostgreSQL database "owenslake" ("springs" schema) at
 db.airsci.com.
}
\details{
@param data Dataframe. Data to be written.
 @param tble Character string. Table in the "springs" schema in the
   "owenslake" database to be wrtitten to.
}

